refactor(lib:design-system): :recycle: 1514 - move into indicators module
