+++ apps/cde-ui/src/app/pages/create-visualization-page/create-visualization-page.component.ts
@@ -19,9 +19,9 @@ import { CsvFileLoaderOptions, CsvFileLoaderService } from '@hra-ui/common/fs';
 import { ButtonsModule } from '@hra-ui/design-system/buttons';
 import { BreadcrumbsComponent } from '@hra-ui/design-system/buttons/breadcrumbs';
 import { ErrorIndicatorComponent } from '@hra-ui/design-system/error-indicator';
+import { StepIndicatorComponent } from '@hra-ui/design-system/indicators/step-indicator';
 import { NavHeaderComponent } from '@hra-ui/design-system/nav-header';
 import { NavigationModule } from '@hra-ui/design-system/navigation';
-import { StepIndicatorComponent } from '@hra-ui/design-system/step-indicator';
 import { TooltipCardComponent, TooltipContent } from '@hra-ui/design-system/tooltip-card';
 import { DeprecatedWorkflowCardComponent } from '@hra-ui/design-system/workflow-card';
 import { DEFAULT_NODE_TARGET_SELECTOR } from '@hra-ui/node-dist-vis';
