+++ libs/design-system/indicators/step-indicator/ng-package.json
