+++ libs/design-system/indicators/step-indicator/src/index.ts
