+++ libs/design-system/indicators/step-indicator/src/lib/step-indicator.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../libs/design-system/styles/vars' as vars;
+@use '../../../../../../libs/design-system/styles/vars' as vars;
 
 :host {
   .step-number {
