+++ libs/design-system/indicators/step-indicator/src/lib/step-indicator.component.ts
